# docs/swagger.yaml
definitions:
...
  models.ColegiosReport:
    properties:
      codigo:
        type: string
      id:
        type: integer
      nombre:
        type: string
    type: object
...
  models.CreateIngresoReq:
    properties:
      activo:
        type: integer
      author:
        type: string
      company_id:
        type: integer
      createdDate:
        type: string
      curso_id:
        type: integer
      fecha:
        type: string
      fechainicial:
        type: string
      fpago:
        type: string
      id:
        type: string
      identificador:
        type: string
      monto:
        type: number
      nrocuotas:
        type: integer
      reserva:
        type: integer
      rutalum:
        type: string
      rutapo:
        type: string
      sale_id:
        type: integer
      status_pago:
        type: string
      tipocomp:
        type: string
      token_flow:
        type: string
      updatedDate:
        type: string
      valorcuota:
        type: number
    type: object
...
  models.CreatePaymentReq:
    properties:
      amount:
        type: number
      auth_code:
        type: string
      auth_date:
        type: string
      author:
        type: string
      card_number:
        type: string
      company_id:
        type: integer
      createdDate:
        type: string
      id:
        type: string
      identifier:
        type: string
      notes:
        type: string
      passenger_id:
        type: integer
      payment_date:
        type: string
      payment_method:
        type: string
      payment_token:
        type: string
      sale_id:
        type: integer
      state:
        type: string
      transaction_ref:
        type: string
      transaction_type:
        type: string
      updatedDate:
        type: string
    type: object
...
  models.CursoReport:
    properties:
      fechanac:
        type: string
      nombrealumno:
        type: string
      nombreapod:
        type: string
      pasaporte:
        type: string
      rutalumno:
        type: string
      rutapod:
        type: string
    type: object
...
  models.IngresoResp:
    properties:
      activo:
        type: integer
      author:
        type: string
      company_id:
        type: integer
      createdDate:
        type: string
      curso_id:
        type: integer
      fecha:
        type: string
      fechainicial:
        type: string
      fpago:
        type: string
      id:
        type: string
      identificador:
        type: string
      monto:
        type: number
      nrocuotas:
        type: integer
      reserva:
        type: integer
      rutalum:
        type: string
      rutapo:
        type: string
      sale_id:
        type: integer
      status_pago:
        type: string
      tipocomp:
        type: string
      token_flow:
        type: string
      updatedDate:
        type: string
      valorcuota:
        type: number
    type: object
...
  models.InitMPagoPaymentReq:
    properties:
      company_id:
        type: integer
      curso_id:
        type: integer
      fechainicial:
        type: string
      identificador:
        type: string
      mpagar:
        type: integer
      nrocuotas:
        type: integer
      sale_id:
        type: integer
      user_rut:
        type: string
      valorcuota:
        type: integer
    type: object
  models.InitMPagoResp:
    properties:
      id_ingreso:
        type: string
      identificador:
        type: string
      monto_pagado:
        type: integer
      preference:
        type: string
      public_key:
        type: string
    type: object
  models.InstallmentResp:
    properties:
      amount:
        type: number
      balance:
        type: number
      company_id:
        type: integer
      createdDate:
        type: string
      curso:
        $ref: '#/definitions/models.CursoReport'
      due_date:
        type: string
      id:
        type: string
      paid_amount:
        type: number
      passenger_id:
        type: integer
      quota_number:
        type: integer
      sale:
        $ref: '#/definitions/models.SaleCursoReport'
      sale_id:
        type: integer
      status:
        type: string
      updatedDate:
        type: string
    type: object
...
  models.MPagoWebhookReq:
    properties:
      action:
        type: string
      data:
        properties:
          id:
            type: string
        type: object
      type:
        type: string
    type: object
...
  models.PaymentResp:
    properties:
      amount:
        type: number
      auth_code:
        type: string
      auth_date:
        type: string
      author:
        type: string
      card_number:
        type: string
      company_id:
        type: integer
      createdDate:
        type: string
      curso:
        $ref: '#/definitions/models.CursoReport'
      id:
        type: string
      identifier:
        type: string
      notes:
        type: string
      passenger_id:
        type: integer
      payment_date:
        type: string
      payment_method:
        type: string
      payment_token:
        type: string
      sale:
        $ref: '#/definitions/models.SaleCursoReport'
      sale_id:
        type: integer
      state:
        type: string
      transaction_ref:
        type: string
      transaction_type:
        type: string
      updatedDate:
        type: string
    type: object
...
  models.SaleCursoReport:
    properties:
      curso:
        type: integer
      establecimiento:
        $ref: '#/definitions/models.ColegiosReport'
      establecimiento_id:
        type: string
      id:
        type: string
      idcurso:
        type: string
      identificador:
        type: string
      program_id:
        type: string
      seller_id:
        type: string
      type_sale:
        type: string
    type: object
  models.SaleResp:
    properties:
      accesscode:
        type: string
      activo:
        type: integer
      author:
        type: string
      comision:
        type: number
      comision_pagada:
        type: integer
      company_id:
        type: integer
      correo_encargado:
        type: string
      createdDate:
        type: string
      cuotas:
        type: integer
      curso:
        type: integer
      descm:
        type: integer
      descrition:
        type: string
      encargado:
        type: string
      establecimiento_id:
        type: integer
      fecha:
        type: string
      fecha_cierre:
        type: string
      fecha_ultpag:
        type: string
      fechacuota:
        type: string
      fechasalida:
        type: string
      id:
        type: string
      idcurso:
        type: string
      identificador:
        type: string
      liberados:
        type: integer
      nroalumno:
        type: integer
      obs:
        type: string
      password:
        type: string
      program_id:
        type: integer
      seller_id:
        type: integer
      sendemail:
        type: integer
      state:
        type: string
      subtotal:
        type: integer
      tipocambio:
        type: number
      type_sale:
        type: string
      updatedDate:
        type: string
      vprograma:
        type: integer
    type: object
...
  models.UpdateInstallmentReq:
    properties:
      amount:
        type: number
      balance:
        type: number
      company_id:
        type: integer
      createdDate:
        type: string
      due_date:
        type: string
      paid_amount:
        type: number
      passenger_id:
        type: integer
      quota_number:
        type: integer
      sale_id:
        type: integer
      status:
        type: string
      updatedDate:
        type: string
    type: object
...
info:
  contact: {}
  description: Powered by scv-go-tools - https://github.com/sergicanet9/scv-go-tools
  title: Go Hexagonal API
paths:
...
  /api/v3.5/ingreso:
    get:
      description: Gets all the ingreso
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SaleResp'
            type: array
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Get inf all ingreso
      tags:
      - ingreso
    post:
      description: Creates a new ingreso
      parameters:
      - description: New colegio to be created
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateIngresoReq'
      responses:
        "201":
          description: OK
          schema:
            $ref: '#/definitions/models.IngresoResp'
        "400":
          description: Bad Request
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Create ingreso
      tags:
      - ingreso
...
  /api/v3.5/iniciopagomp:
    post:
      description: Inicializa el pago a través de Mercado Pago y devuelve los datos
        de preferencia y clave pública
      parameters:
      - description: Configuración inicial del pago
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.InitMPagoPaymentReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.InitMPagoResp'
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Init Mercado Pago Payment
      tags:
      - mpago
  /api/v3.5/installment:
    get:
      description: Gets all the installment
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.InstallmentResp'
            type: array
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Get all installment
      tags:
      - installment
    post:
      description: Creates a new payment
      parameters:
      - description: New payment to be created
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreatePaymentReq'
      responses:
        "201":
          description: OK
          schema:
            $ref: '#/definitions/models.PaymentResp'
        "400":
          description: Bad Request
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Create payment
      tags:
      - payment
  /api/v3.5/installment/{id}:
    delete:
      description: Delete a installment
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Delete installment
      tags:
      - installment
    get:
      description: Gets a installment by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.InstallmentResp'
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Get installment by ID
      tags:
      - installment
    patch:
      description: Updates a installment
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: installment
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/models.UpdateInstallmentReq'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Update installment
      tags:
      - installment
...
  /api/v3.5/mpago/verificar:
    get:
      description: Recibe el retorno del usuario desde Mercado Pago, verifica el pago
        y redirige a la pantalla de resultado
      parameters:
      - description: ID del pago en Mercado Pago
        in: query
        name: payment_id
        type: string
      - description: ID de la preferencia en Mercado Pago
        in: query
        name: preference_id
        type: string
      responses:
        "302":
          description: Redirección a /mpagopagos/resultado
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: object
      security:
      - Bearer: []
      summary: Retorno y verificación Mercado Pago
      tags:
      - mpago
  /api/v3.5/mpago/webhook:
    post:
      consumes:
      - application/json
      description: Procesa las notificaciones webhook enviadas por Mercado Pago tras
        eventos de pago
      parameters:
      - description: Datos de la notificación webhook
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.MPagoWebhookReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Webhook Mercado Pago
      tags:
      - mpago
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"